added while loop & the final console.log statements
I chose to leave the starCheck (while loop) out of the level object and
make it its own stand alone function. The function checks to see if the
player has enough stars to move onto the next cave and the level object
is meant for the individual levels in each cave.